Add files via upload
# Mini IMDB-clone

**Following are the Features of this webapp:-**

**Home Page:**

Through this page, Users can search for any Movie or TV series. We have used the OMDB API. When the user types in the search box, real-time search suggestions are displayed, similar to Google. Each search result will have a "Favourite" button, allowing users to add the movie to their "My Favourite Movies" list. Clicking on a particular search result will open a new page with more information about the selected movie.

**My Favourite Movies Page:** 

By means of this page users can get UI to access and manipulate their favourite movies. Users can view a list of all their favourite movies on this page. The list of favourite movies will be made persistent, meaning it will be retained even after closing or refreshing the browser. Each movie in the favourites list will have a "Remove from favourites" button, allowing users to remove a movie from the list.

# Short description of this App:

The app has been implemented using vanilla JavaScript, and it will interact with the OMDB API to fetch movie data. The user interface will be designed using CSS, and it uses a CSS framework like Bootstrap to assist with styling as desired. The app's data, including the list of favourite movies, will be stored in the browser's local storage to make it persistent across sessions.